techservice/requirements.txt
asgiref==3.7.2
Django==3.2.9
python-dotenv==1.0.0
sqlparse==0.4.4
Werkzeug==2.3.7
Pillow==8.4.0
openpyxl==3.1.5
django-ckeditor==6.7.2
django-jazzmin==2.6.0
python-decouple==3.8
django-js-asset==2.0.0
django-widget-tweaks==1.5.0
gunicorn==20.1.0
